// angular/angular-template/src/app/model/news.ts
export class News {
    newsId: string;
    newsName: string;
    description: string;
    content: string;
    hightlight: string;
    image: string;
    createDate: Date;
    isDelete: boolean;
    employeeId: string;
  
    // constructor(  newsId: string,
    //     newsName: string,
    //     description: string,
    //     content: string,
    //     hightlight: string,
    //     image: string,
    //     createDate: Date,
    //     isDelete: boolean,
    //     employeeId: string) {
    //   this.newsId = newsId;
    //   this.newsName = newsName;
    //   this.description = description;
    //   this.content = content;
    //   this.hightlight = hightlight;
    //   this.image = image;
    //   this.createDate = createDate;
    //   this.isDelete = isDelete;
    //   this.employeeId = employeeId;
    // }
  }
  